+++ src/Pages/style/ArticleTag.module.css
@@ -19,23 +19,19 @@
     margin-top: 10%;
 }
 
-.tagsContainer {
+.searchResult {
+    margin-top: 5%;
+    width: 100%;
+    background-color: white;
+    border-radius: 10px;
+    padding: 10px;
+}
+
+.searchResults {
     display: flex;
     flex-wrap: wrap;
     justify-content: center;
-    align-items: center;
-    margin-top: 5%;
-    margin-bottom: 5%;
-    width: 60%;
-}
-
-.tag {
-    margin: 5px;
-    padding: 5px;
-    border-radius: 5px;
-    background-color: #d9d9d9;
-    color: black;
-    cursor: pointer;
+    
 }
 
 .footerContainer {
